Store and pass variant data in the Identity's authentication. (#3788)
This is preparing the introduction of an authentication SPI. One prerequisite for supporting different authorization mechanisms is the ability to pass the information contained in this authorization artifacts down to the user containers (e.g. to call out to other services using these artifacts). This ability is implemented by serializing the information into a JSON structure
that is mixed in into environment available to the runtimes.