add /numHealthyInvokers route to controller
Add a route to easily get a count of healthy invokers. This route can
be used to as a check for the system being ready to actually execute
actions (for example as a readiness probe in a Helm chart or a
TravisCI built script). Although this information could be extracted by
client-side processing of the result of the /invokers route, it is
convenient to do the math on the server side and simply return the count.