Properly consume and optimize reading of http entities, drop connection reuse.
- Explicitly consume and close the response's entity in any case (even if ignored) to make sure the connections get released properly and are not leaked.
- Use optimized path for consuming the entire entity into a string if its length is within bounds.
- Don't reuse connections.

To the last point: Reusing connections when a runtime doesn't support it adds a significant latency overhead when closing the response's entity. That's likely due to some sort of mismatched behavior (client wants to keep the connection open, the server doesn't even know the concept). The latency overhead in those cases (10-20ms of latency added in latency tests) seem far higher than the latency overhead of establishing a new connection (not measurable in latency tests).
Dropping connection reuse also solves any issues that might be encountered due to the pause/resume cycles of the container, in which sockets don't react to any event at all.