Implement full specification of docker image names.
Today we have a very basic way of parsing docker image names, which even prohibit the use of ports in the domain names and/or providing a fully-qualified digest to pull.

This parser implements the full specification of docker image names to allow OpenWhisk to take all kinds of image references possible.