Fix max.poll.interval.ms setting of KafkaConsumer.
The `max.poll.interval.ms` setting needs to be high enough to not be thrown off by all actions in the system consuming the maximum duration of their respective action timeout, ideally also adding enough slack to account for image pulls taking long, or docker runs being very slow.

The name of the setting was also misspelled so we've been running on the default.