Synchronize access to the KafkaConsumer.
The KafkaConsumer is **not** thread-safe, thus concurrent access to it needs to be avoided.

Per https://twitter.github.io/scala_school/concurrency.html, a synchronized block should be quite cheap as long as there is low contention. The contention in this case is controllable via the cadence the monitoring code is running (default 60 seconds). The cost of this synchronization thus should be relatively low.