Bump Kafka's fetch.max.wait duration and adjust poll's timeout.
Kafka's `fetch.max.wait` duration determines how long the server will wait to send a response if *no record is present*.
The timeout set on `poll` determines how long the client blocks if *no record is present*.

Note that both of these only block if no records are present. Latency should not be impacted, if records are arriving quickly, as both of these timeouts will not apply if records are already there.

The `fetch.max.wait` duration determines how often Kafka is queried though. Big systems can slow down Kafka even by the amount of polling that is happening, even if all the topics are empty.