Adjust Kafka's config to not contain any more warnings.
Currently, the KafkaConsumer warns about a misconfiguration because we added a fantasy value to the kafka config itself. There is a value to control the window size of metrics reporting which roughly fits the needs of the value we need, thus switching to that value also gets rid of the warning in the log.