Support for start-of-activation log sentinel
Add support for filtering out an optional start-of-activation log sentinel
line that will contain the activation id in addition to the sentinel marker.
For runtimes that inject this sentinel line into their outputs, it becomes
possible to do an asynchronous and streaming operation on their log streams
to forward log lines enriched with activation ids to a platform logging service.

Also include a Dockerfile defining a lightweight log processing agent
using fluent-bit to demonstrate how the logs could be processed.