Don't lookup config on every container-creation.
On profiling the invoker I saw, that the lookup of the config blocks the invoker-threads on the beginning of a burst.
The reason is, that the config is read on every container-creation. It is enough to read this configuration once.
I also verified, that the threads are not blocked after the change.