Switch from alpine to jessie-slim for runner utility images
The Alpine based images have a nasty problem with DNS failures that
tends to surface when running them in Kubernetes.  After a fair amount
of poking around, it seems like the only reliable fix is to not use
Alpine images on Kubernetes until upstream bug fixes in various layers
of the software stack, including the Linux kernel propagate to the
Alpine releases.  For more context,
see:
  https://github.com/gliderlabs/docker-alpine/issues/255
  https://github.com/kubernetes/kubernetes/issues/56903
  https://www.weave.works/blog/racy-conntrack-and-dns-lookup-timeouts