Fix controller logs in docker-machine
With the change https://github.com/apache/incubator-openwhisk/pull/3579, controller logs file was not being created due to lack of permissions when OW is deployed on docker-machine. This is because ansible is not able to create a folder with full permissions within a mounted folder in docker-machine. However creating a file with full permissions through ansible does work.