Diagnostic info and metrics for Docker command failures and timeouts (#4070)
This change improves diagnostic information for failing Docker commands as well as timed out Docker commands:

* For all failures (including timeouts), a textual representation for exit status values is logged.
* Timeouts are explicitly detected and reported using a specialized exception allowing for a better timeout handling on higher layers of the Docker container implementation.
* Emit counter metric on Docker command timeout

This change introduces a set of new counter metrics that are emitted if a Docker command is terminated because of a timeout. A high number of such timeout occurrences is usually an indication for highly loaded invokers. The new metrics help to identify such invokers.