Make ContainerProxyTests more stable. (#4289)
As the active ack is sent asynchronously in the invoker, it could happen, that the order of them is wrong in the ContainerProxyTest. This has been proven with the following debug information: https://github.com/apache/incubator-openwhisk/pull/4278

This PR makes these tests more stable by not relying on the order of active acks.