Make WskRestRuleTests more stable.
If there is an deployment with several controllers, the test `Whisk rules should preserve rule status when a rule is updated` fails intermittently. The reason is, that there are several modifications on the rule. These modifications are already retried. But enabling and disabling the rule also modifies the database document. But that change is currently not retried.
This PR puts the enabling/disabling of the rule into the retry block as well to make this test more stable against our eventual consistency.

Co-authored-by: Somaya Jamil <jamilsom@de.ibm.com>