Partition action arguments into env vars and main args.
Activation arguments are partitioned so that there is an environment variables set and the rest are "main" args.
The predicate for partioning is the property must start with a capital letter and have a value of type string|number|boolean.

Update action proxy invoke helper to accept environment variables to export to action.