Add more diagnostic information to completion ack processing
Forced completion acks can be a source of overloaded invokers because forced acks free up invoker slots in the load balancer. If the invoker is just "late" with running an activation so that the invoker sends the result / completion ack after the completion ack timeout has been reached in the load balancer, the load balancer may send new activations to the invoker while it is still fully occupied. As a result, these new activations have to wait for a container and can cause forced completion acks in turn. This mechanism can lead to a never-ending chain reaction if an invoker has a constant stream of workload.

Goal of this change is to improve visibility and diagnostic information so that this mechanism can be better understood and fixed in the next step.

* Add metrics for forced completion acks.
* Log more diagnostic information when forcing a completion ack due to timeout.
* Log a warning if a completion ack arrives after it has already been forced - this is an indication that action processing took too long. Today, this situation is only logged as debug message.
* Improve code comments.
* Provide a function for calculating the completion ack timeout. In the long term, this should be unified with the action wait timeout in sequence activations.