Allow parameters to be designated as init time properties (#4559)
Allows parameters to be designated as init time parameters. The result of such a designation is that the corresponding parameters are communicated to the action at initialization time and should be exported to the action's environment accordingly. This change requires updates to the action runtimes for the latter details and will be done separately.

The env values are string representation of the actual JSON parameter value. So if the parameter value is an object then env value would be the json object's string representation. 

If the runtime supports this feature then it must implement the `BasicActionRunnerTests#testEnvPartition` test