GoLang docker image: actionloop-golang ==> action-golang
openwhisk-runtime-go PR#109 changed the name used to publish docker
images from actionloop-golang-VERSION to action-golang-VERSION. This
PR updates the core system runtime.json and documentation to use the
new names. All old images are still available on dockerhub and will
not be removed or renamed, so this should be a non-disruptive change.