remove experimental KubernetesClientWithInvokerAgent (#4785)
Remove experimental KubernetesClientWithInvokerAgent implementation of
the ContainerFactory SPI to simplify the codebase.  This code was an
exploration of implementing pause/unpause functionality when using the
KubernetesContainerFactory, but was never implemented to a useful
level of functionality.  Most seriously, if an invoker or invoker
agent crashed it would leave orphaned suspended pods that Kubernetes
was not able to cleanly remove without cluster admin-level
intervention since they had been paused "behind the back" of the
Kubernetes scheduler using low-level mechanisms.  A secondary
consideration is that the actual invoker agent (go code in the
deploy-kube project) was only implemented for the case when the
underlying container engine was Docker. Given the subsequent shift
away from Docker to containerd or cri-o, the invoker agent did not
support recent versions of Kubernetes or OpenShift.