Add single entrypoint proxy interface documentation
Several implementations of a single entrypoint proxy contract exist
for multiple different runtimes. In effort to document this existing
code, these changes have been put down as part of the documentation
for OpenWhisk since they do not properly fit under the definition of
any one runtime. In addition, a small footer has been added to the
docs/action-new.md to make note of and link to the
docs/single_entrypoint_proxy_contract.md describing and giving examples
of this contract.