Add single entrypoint proxy interface documentation (#4850)
* Add single entrypoint proxy interface documentation

Several implementations of a single entrypoint proxy contract exist
for multiple different runtimes. In effort to document this existing
code, these changes have been put down as part of the documentation
for OpenWhisk since they do not properly fit under the definition of
any one runtime. In addition, a small footer has been added to the
docs/action-new.md to make note of and link to the
docs/single_entrypoint_proxy_contract.md describing and giving examples
of this contract.

* Add apache license to docs/single_entrypoint_proxy_contract.md

* Convert html to markdown table in single entrypoint execution environment docs

* Table experiment part 2

* Table experiment part 3

* Table experiment part 4

* Table experiment part 5

* Revert "Html to markdown conversion commits"

This reverts commit 188d442c1a33513f91be349a897b6798cdd9852c.
This reverts commit 7107927c420d5b5ffc092f296daf557712ced51e.
This reverts commit 3f21577af72d2debad7d5b7c9c8f8af7609ef152.
This reverts commit c58ca4cff1585fa3ad2f3ce3c9cf3fc034c0b35b.
This reverts commit f87c08bd3279c3b0c7e618f94ac6a31cfec8aee3.

* Add table with links to NodeJS tests to single entrypoint docs

* Reword additional execution environments section in action-new doc

* Add missing word to additional execution environment in action-new doc

* Add a comment to the nodejs table in single entrypoint docs