Remove `fail-fast` matrix strategy from GitHub Actions (#5421)
Currently, every GitHub Actions workflow contains the `fail-fast` matrix strategy. However, none of these workflows are actually utilizing a matrix. Further, the on demand workflow has an improper indentation of this strategy block that causes a syntactical error.

This change remove the no-op strategy altogether to fix the broken on demand workflow and remove the IDE warnings regarding the missing matrix declaration.